+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ Changelog
 
 Version 0.8-dev
 ---------------
-| **Date:** November 10, 2021
+| **Date:** December 1, 2021
 | **Release:** v0.8-dev
 | **Version of CHAOS:** CHAOS-7.8 (0708)
 
@@ -11,6 +11,10 @@ Features
 ^^^^^^^^
 * Added :func:`chaosmagpy.coordinate_utils.sh_analysis`, which performs a
   spherical harmonic expansion.
+
+Bugfixes
+^^^^^^^^
+* Fixed numpy broadcasting error in :func:`chasomagpy.data_utils.mjd2000`.
 
 Version 0.7.1
 -------------